* [x] 1. Use bytes.IndexByte instead of bytes.Split.
* [ ] 2. Unify bytes.Buffer. You can build outer header in-place.
* [x] 3. Cache buffer in RLogger. Guard it with mutex.

see also #2