@hhatto 2016-07-20
now

```
benchmark            old ns/op     new ns/op     delta
BenchmarkWrite-4     2721          1033          -62.04%

benchmark            old allocs     new allocs     delta
BenchmarkWrite-4     18             1              -94.44%

benchmark            old bytes     new bytes     delta
BenchmarkWrite-4     1072          32            -97.01%
```