@hhatto 2016-07-21
final result

```
benchmark            old ns/op     new ns/op     delta
BenchmarkWrite-4     2721          953           -64.98%

benchmark            old allocs     new allocs     delta
BenchmarkWrite-4     18             0              -100.00%

benchmark            old bytes     new bytes     delta
BenchmarkWrite-4     1072          0             -100.00%
```